+++ device-management-application/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [SW-14061] et-powertop: Added support for displaying ASIC voltages.
 - [SW-14337] dev_mngt_service: Added support for DM_CMD_RESET_ETSOC.
 - [SW-14586] merge all trace operations into one option.
+- [SW-14616] dev_mngt_service: Added option to print PCI_SLOT_NAME.
 ### Changed
 - [SW-13832] Use DM_CMD_GET_MM_STATS instead of getTraceBufferServiceProcessor
 - Changed other Power from 5.75W to 3.75W